+++ pkg/manifests/manifests_test.go
@@ -271,7 +271,7 @@ func TestUnconfiguredManifests(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	_, err = f.KubeStateMetricsDeployment()
+	_, err = f.KubeStateMetricsDeployment(false)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -2638,9 +2638,7 @@ k8sPrometheusAdapter:
 				t.Fatal(err)
 			}
 
-			f := NewFactory("openshift-monitoring", "openshift-user-workload-monitoring",
-				c, defaultInfrastructureReader(), &fakeProxyReader{},
-				NewAssets(assetsPath), &APIServerConfig{}, &configv1.Console{})
+			f := NewFactory("openshift-monitoring", "openshift-user-workload-monitoring", c, defaultInfrastructureReader(), &fakeProxyReader{}, NewAssets(assetsPath), &APIServerConfig{}, &configv1.Console{})
 
 			d, err := f.PrometheusAdapterDeployment("foo", map[string]string{
 				"requestheader-allowed-names":        "",
@@ -3724,13 +3722,14 @@ func TestKubeStateMetrics(t *testing.T) {
 
 	f := NewFactory("openshift-monitoring", "openshift-user-workload-monitoring", c, defaultInfrastructureReader(), &fakeProxyReader{}, NewAssets(assetsPath), &APIServerConfig{}, &configv1.Console{})
 
-	d, err := f.KubeStateMetricsDeployment()
+	d, err := f.KubeStateMetricsDeployment(true)
 	if err != nil {
 		t.Fatal(err)
 	}
 
 	kubeRbacProxyTLSCipherSuitesArg := ""
 	kubeRbacProxyMinTLSVersionArg := ""
+	gotKubeStateMetricsCustomResourceStateConfigFile := ""
 	for _, container := range d.Spec.Template.Spec.Containers {
 		switch container.Name {
 		case "kube-state-metrics":
@@ -3740,6 +3739,8 @@ func TestKubeStateMetrics(t *testing.T) {
 			if !reflect.DeepEqual(container.Resources, *f.config.ClusterMonitoringConfiguration.KubeStateMetricsConfig.Resources) {
 				t.Fatal("kube-state-metrics resources incorrectly configured")
 			}
+
+			gotKubeStateMetricsCustomResourceStateConfigFile = getContainerArgValue(d.Spec.Template.Spec.Containers, kubeStateMetricsCustomResourceStateConfigFileFlag, container.Name)
 		case "kube-rbac-proxy-self", "kube-rbac-proxy-main":
 			if container.Image != "docker.io/openshift/origin-kube-rbac-proxy:latest" {
 				t.Fatalf("%s image incorrectly configured", container.Name)
@@ -3772,7 +3773,11 @@ func TestKubeStateMetrics(t *testing.T) {
 		t.Fatal("kube-state-metrics topology spread constraints WhenUnsatisfiable not configured correctly")
 	}
 
-	d2, err := f.KubeStateMetricsDeployment()
+	if gotKubeStateMetricsCustomResourceStateConfigFile != kubeStateMetricsCustomResourceStateConfigFileFlag+kubeStateMetricsCustomResourceStateConfigFile {
+		t.Fatal("expected kube-state-metrics to enable custom-resource-state metrics")
+	}
+
+	d2, err := f.KubeStateMetricsDeployment(true)
 	if err != nil {
 		t.Fatal(err)
 	}